#49 Improve log output

+++ resources/remove-user.sh
@@ -5,9 +5,21 @@ set -o pipefail
 
 USERNAME="${1}"
 
-RAILS_ENV=production bundle exec rails console <<< "
+OUTPUT="$(RAILS_ENV=production bundle exec rails console <<< "
 user = User.where(login: '${USERNAME}').first
-unless user.nil?
-  user.destroy
+begin
+  user.destroy unless user.nil?
+rescue => error
+  puts error.message
+  raise 'User was not saved'
 end
-" >> /dev/null
+")"
+
+# There is no way to get an exit code on error. So we check if exception raised text appears and exit manually.
+if [[ "${OUTPUT}" == *"RuntimeError (User was not saved)"* ]]; then
+  echo "Could not remove user ${USERNAME} due to error: "
+  printf '%s\n' "${OUTPUT#*end}"
+  exit 1
+else
+  echo "Removed user ${USERNAME} successfully."
+fi
